requirements/development.txt
# Include base requirements
-r base.txt

# Development Tools
django-debug-toolbar>=4.1.0,<5.0.0
django-extensions>=3.2.1,<4.0.0
ipython>=8.13.0,<9.0.0
werkzeug>=2.3.4,<3.0.0

# Testing
pytest>=7.3.1,<8.0.0
pytest-django>=4.5.2,<5.0.0
factory-boy>=3.2.1,<4.0.0
faker>=18.9.0,<19.0.0
coverage>=7.2.5,<8.0.0

# Code Quality
black>=23.3.0,<24.0.0
isort>=5.12.0,<6.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.3.0,<2.0.0
django-stubs>=4.2.0,<5.0.0
djangorestframework-stubs>=3.14.0,<4.0.0

# Documentation
mkdocs>=1.4.3,<2.0.0
mkdocs-material>=9.1.11,<10.0.0
